feat(pokemon): DJ-30 implement CreatePokemon use case
This commit introduces the CreatePokemon use case to decouple
the creation logic of Pokémon from the controller. By creating a
separate use case, we enhance the modularity and testability
of our application. Controllers can now focus solely on handling
HTTP requests and responses, while the use case encapsulates the
business logic for creating a Pokémon.

The CreatePokemonController now receives a CreatePokemon function
instead of a repository directly. This change allows for
better separation of concerns and paves the way for easier
mocking during tests. The setup function for the use case
is defined in `src/domain/use-cases/create-pokemon.ts`, which
takes a Pokémon repository and returns an asynchronous function.

Additionally, the factory function for CreatePokemonController
has been updated to set up the new use case, thereby ensuring
that the controller operates with the latest architecture.
Existing tests for CreatePokemonController have also been adapted
to reflect these changes, focusing on testing the use case
instead of the repository directly.