feat(battle-pokemon): DJ-19 create controller for Pokémon battles
This commit introduces the `BattlePokemonController` to handle
battles between Pokémon entities. The controller is set up by
injecting the battle use case, `setupBattlePokemon`, with the
necessary repository implementation, `makePgPokemonRepository`.

The motivation for adding this controller lies in the need to
provide a structured way to manage battles in the application,
ensuring that the business logic for battling Pokémon is encapsulated
and can be reused across different routes.

A new route has been established that maps to this controller,
allowing for POST requests to initiate battles between two Pokémon
identified by their respective IDs. This functionality expands the
existing API capabilities and enhances the interactive features
for users.

Additionally, the controller has been integrated into the main
exports of the controller module for easy access within the
application structure. This setup will facilitate future
developments related to battle mechanics and Pokémon interactions.